pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sanger-qc-trim"
version = "0.1.0"
description = "QC and trimming tool for Sanger sequencing reads (.ab1 and .phd.1 files)"
authors = [{name = "JGI Science", email = "noreply@jgi.doe.gov"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "biopython>=1.79",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "pyarrow>=6.0.0",
    "typer>=0.9.0",
    "tqdm>=4.62.0",
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
sangerqc = "sanger_qc_trim.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["sanger_qc_trim*"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"
